Move raise_on_call_returned_empty to proxies utils
